+++ src/features/preferences/userDefaults.ts
@@ -11,6 +11,7 @@ import { IPreferenceDefault } from "@/features/preferences/userPreferences.types
 export const preferenceDefaults: IPreferenceDefault = {
 	defaultEmpireUuid: undefined,
 	defaultCXUuid: undefined,
+	defaultBuyItemsFromCX: true,
 	burnDaysRed: 5,
 	burnDaysYellow: 10,
 	burnResupplyDays: 20,